requirements.txt
fastapi==0.111.0

SQLAlchemy==2.0.30
PyMySQL==1.1.1
alembic==1.13.1
pydantic-settings==2.6.0

bcrypt==4.1.3

loguru==0.7.2

git+https://github.com/ajrlewis/webkit.git
git+https://github.com/ajrlewis/huggingkit.git
git+https://github.com/ajrlewis/imagekit.git
